Get PolyWrapper from PolyBridge in update-monitor.
